requirements.txt
bleach==6.2.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
python-dotenv==1.1.1
python-slugify==8.0.4
text-unidecode==1.3
webencodings==0.5.1
Werkzeug==3.1.3
